_mark_down_compupdated

+++ src/comps/submission/new/panels/Attributes.jsx
@@ -7,32 +7,32 @@ import { findAndInsertTree, findChildrenByPath, findPath, deleteByPath, checkPat
 
 export function AttributesTab({submission, setSubmission, setComponentKey, componentKey }) {
     
-    /**
-     * @description Handle dataset selection 
-     * @param {String} attribute_tag - Historically - should be removed. 
-     * @param {String} trait - The actual trait that was selected by its tag. 
-     */
-    const handleTraitSelection = (attribute_tag, trait_tag) => {
-        //const attribute_tag = trait.attribute_tag 
-        let selected_traits =  submission.selected_traits.slice() //mission.selected_traits
+    // /**
+    //  * @description Handle dataset selection 
+    //  * @param {String} attribute_tag - Historically - should be removed. 
+    //  * @param {String} trait - The actual trait that was selected by its tag. 
+    //  */
+    // const handleTraitSelection = (attribute_tag, trait_tag) => {
+    //     //const attribute_tag = trait.attribute_tag 
+    //     let selected_traits =  submission.selected_traits.slice() //mission.selected_traits
         
-        if (!_.has(selected_traits, attribute_tag)) {
-            selected_traits[attribute_tag] = [trait_tag]
-        }
-        else {
+    //     if (!_.has(selected_traits, attribute_tag)) {
+    //         selected_traits[attribute_tag] = [trait_tag]
+    //     }
+    //     else {
             
-            selected_traits[attribute_tag] = addStringToArrayOrRemove({ array: selected_traits[attribute_tag], string: trait_tag })
-            if (selected_traits[attribute_tag].length === 0) {
-                delete selected_traits[attribute_tag]
-            }
-        }
-        setSubmission(prevValues => { return {...prevValues, selected_traits}})
-    }
+    //         selected_traits[attribute_tag] = addStringToArrayOrRemove({ array: selected_traits[attribute_tag], string: trait_tag })
+    //         if (selected_traits[attribute_tag].length === 0) {
+    //             delete selected_traits[attribute_tag]
+    //         }
+    //     }
+    //     setSubmission(prevValues => { return {...prevValues, selected_traits}})
+    // }
 
 
     const getSelectionByPath = (path) => {
         const selection = findChildrenByPath(submission.selected_traits, path)
-        console.log(selection, "getSelectionByPath", submission.selected_traits, path)
+        // console.log(selection, "getSelectionByPath", submission.selected_traits, path)
         return selection
     }
 
@@ -42,15 +42,18 @@ export function AttributesTab({submission, setSubmission, setComponentKey, compo
         return selection
     }
 
-    const handleTraitSelectionHierarchy = (path, _info) => {
+    const handleTraitSelectionHierarchy = (path, row_idcs, single_child_level, single_child_type) => {
+        console.log(row_idcs, single_child_level, single_child_type, "info")
         let selected_traits = submission.selected_traits.slice() //mission.selected_traits
         //check if path exists in the selected traits
         const pathExists = checkPathExists(selected_traits, path)
+        const isValueInput = _.last(path).type === "trait" && _.has(_.last(path), "value") && _.isString(_.last(path).value)
         // if it exists, delete it, if not, add it.
-        if (pathExists) {
+        if (pathExists && !isValueInput) {
             deleteByPath(selected_traits, path)
         }
         else {
+            
             findAndInsertTree(selected_traits, path)
         }
         setSubmission(prevValues => { return {...prevValues, selected_traits}})
@@ -78,7 +81,7 @@ export function AttributesTab({submission, setSubmission, setComponentKey, compo
             getSelectionByPath={getSelectionByPath}
             submission_tag={submission.tag}
             attributeTraits={submission.selected_traits}
-            onChildrenSelection={handleTraitSelection}
+            onChildrenSelection={handleTraitSelectionHierarchy}
             handleTraitRemove={handleTraitSelectionHierarchy}
          />
         {/* userUnitInput={submission.userUnitInput}
